add confession services

+++ Repositories/IRepository.cs
@@ -4,12 +4,12 @@ namespace Saycret.Repositories
 {
     public interface IRepository<T> where T : class
     {
-        public T GetByID(int id);
+        public T GetByID(long id);
         List<T> FindAll(Expression<Func<T, bool>> expression, int? take, int? skip);
         T Add(T entity);
-        void Delete(int id);
-        T Update(T entity, int id);
-        int CountAll();
-        int CountSpecificItems(Expression<Func<T, bool>> expression);
+        void Delete(long id);
+        T Update(T entity, long id);
+        long CountAll();
+        long CountSpecificItems(Expression<Func<T, bool>> expression);
     }
 }
